+++ src/databaseTests/src/test/kotlin/org/vaccineimpact/api/databaseTests/RepositoryTestSuite.kt
@@ -5,6 +5,8 @@ import org.junit.BeforeClass
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.vaccineimpact.api.databaseTests.tests.*
+import org.vaccineimpact.api.databaseTests.tests.burdenEstimateRepository.UploadBurdenEstimateTests
+import org.vaccineimpact.api.databaseTests.tests.burdenEstimateRepository.BurdenEstimateRepositoryTestsDeprecated
 import org.vaccineimpact.api.databaseTests.tests.modellingGroupRepository.GetModellingGroupTests
 import org.vaccineimpact.api.databaseTests.tests.modellingGroupRepository.GetResponsibilitiesTests
 import org.vaccineimpact.api.databaseTests.tests.modellingGroupRepository.GetResponsibilityCoverageSetsTests
@@ -15,7 +17,7 @@ import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
 // Keep these sorted alphabetically, for consistency
 @RunWith(Suite::class)
 @Suite.SuiteClasses(
-        BurdenEstimateRepositoryTests::class,
+        UploadBurdenEstimateTests::class,
         BurdenEstimateRepositoryTestsDeprecated::class,
         DiseaseTests::class,
         GetDemographicsTests::class,
