+++ src/databaseTests/src/test/kotlin/org/vaccineimpact/api/databaseTests/tests/burdenEstimateRepository/UploadBurdenEstimateTests.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api.databaseTests.tests
+package org.vaccineimpact.api.databaseTests.tests.burdenEstimateRepository
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
@@ -8,12 +8,6 @@ import org.vaccineimpact.api.app.errors.DatabaseContentsError
 import org.vaccineimpact.api.app.errors.InconsistentDataError
 import org.vaccineimpact.api.app.errors.OperationNotAllowedError
 import org.vaccineimpact.api.app.errors.UnknownObjectError
-import org.vaccineimpact.api.app.repositories.BurdenEstimateRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqBurdenEstimateRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqModellingGroupRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqScenarioRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqTouchstoneRepository
-import org.vaccineimpact.api.databaseTests.RepositoryTests
 import org.vaccineimpact.api.db.JooqContext
 import org.vaccineimpact.api.db.Tables.*
 import org.vaccineimpact.api.db.direct.*
@@ -21,45 +15,68 @@ import org.vaccineimpact.api.db.fromJoinPath
 import org.vaccineimpact.api.db.toDecimal
 import org.vaccineimpact.api.models.BurdenEstimate
 import java.math.BigDecimal
-import java.time.LocalDateTime
-import java.time.Month
-import java.time.ZoneOffset
 
-// This class tests a single, deprecated method - 'addBurdenEstimateSet'
-// It can be removed entirely when the method is removed
-class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRepository>()
+class UploadBurdenEstimateTests : BurdenEstimateRepositoryTests()
 {
-    private data class ReturnedIds(val modelVersion: Int?, val responsibility: Int)
-
-    override fun makeRepository(db: JooqContext): BurdenEstimateRepository
+    @Test
+    fun `can create burden estimate set with empty status`()
     {
-        val scenario = JooqScenarioRepository(db.dsl)
-        val touchstone = JooqTouchstoneRepository(db.dsl, scenario)
+        var returnedIds: ReturnedIds? = null
+        var setId: Int? = null
 
-        val modellingGroup = JooqModellingGroupRepository(db.dsl, touchstone, scenario)
-        return JooqBurdenEstimateRepository(db.dsl, scenario, touchstone, modellingGroup)
+        given { db ->
+            returnedIds = setupDatabase(db)
+        } makeTheseChanges { repo ->
+            setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+        } andCheckDatabase { db ->
+            checkBurdenEstimateSetMetadata(db, setId!!, returnedIds!!, "empty")
+        }
     }
 
-    private val scenarioId = "scenario-1"
-    private val groupId = "group-1"
-    private val touchstoneId = "touchstone-1"
-    private val modelId = "model-1"
-    private val modelVersion = "version-1"
-    private val username = "some.user"
-    private val timestamp = LocalDateTime.of(2017, Month.JUNE, 13, 12, 30).toInstant(ZoneOffset.UTC)
-
     @Test
-    fun `can add burden estimate set`()
+    fun `can populate burden estimate set`()
     {
+        var setId: Int? = null
         var returnedIds: ReturnedIds? = null
 
         given { db ->
             returnedIds = setupDatabase(db)
         } makeTheseChanges { repo ->
-            repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+            setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+            repo.populateBurdenEstimateSet(setId!!, groupId, touchstoneId, scenarioId, data)
         } andCheckDatabase { db ->
-            val setId = checkBurdenEstimateSetMetadata(db, returnedIds!!)
-            checkBurdenEstimates(db, setId)
+            checkBurdenEstimates(db, setId!!)
+            checkBurdenEstimateSetMetadata(db, setId!!, returnedIds!!, "complete")
+        }
+    }
+
+
+    @Test
+    fun `cannot populate a set unless status is empty`()
+    {
+        JooqContext().use {
+            val returnedIds = setupDatabase(it)
+            val setId = it.addBurdenEstimateSet(returnedIds.responsibility, returnedIds.modelVersion!!, username, "complete")
+
+            val repo = makeRepository(it)
+            assertThatThrownBy {
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, data)
+            }.isInstanceOf(OperationNotAllowedError::class.java)
+                    .hasMessage("the following problems occurred:\nThis burden estimate set already contains estimates." +
+                            " You must create a new set if you want to upload any new estimates.")
+        }
+    }
+
+
+    @Test
+    fun `populate set throws unknown object error if set does not exist`()
+    {
+        JooqContext().use {
+            setupDatabase(it)
+            val repo = makeRepository(it)
+            assertThatThrownBy {
+                repo.populateBurdenEstimateSet(12, groupId, touchstoneId, scenarioId, data)
+            }.isInstanceOf(UnknownObjectError::class.java)
         }
     }
 
@@ -71,34 +88,37 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
         given { db ->
             returnedIds = setupDatabase(db)
         } makeTheseChanges { repo ->
-            setId = repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+            setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+            repo.populateBurdenEstimateSet(setId!!, groupId, touchstoneId, scenarioId, data)
         } andCheckDatabase { db ->
             checkCurrentBurdenEstimateSet(db, returnedIds!!, setId!!)
         }
     }
 
     @Test
-    fun `cannot add burden estimates with diseases that do not match scenario`()
+    fun `cannot create burden estimates with diseases that do not match scenario`()
     {
         val badData = data.map { it.copy(disease = "YF") }
         JooqContext().use { db ->
             setupDatabase(db)
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, badData, username, timestamp)
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, badData)
             }.isInstanceOf(InconsistentDataError::class.java)
         }
     }
 
     @Test
-    fun `cannot add burden estimates with non-existent countries`()
+    fun `cannot populate burden estimates with non-existent countries`()
     {
         val badData = data.map { it.copy(country = "FAKE") }
         JooqContext().use { db ->
             setupDatabase(db)
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, badData, username, timestamp)
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, badData)
             }.isInstanceOf(UnknownObjectError::class.java).matches {
                 (it as UnknownObjectError).typeName == "country"
             }
@@ -106,38 +126,39 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
     }
 
     @Test
-    fun `cannot add burden estimate set if group has no model`()
+    fun `cannot create burden estimate set if group has no model`()
     {
         JooqContext().use { db ->
             setupDatabase(db, addModel = false)
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+                repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
             }.isInstanceOf(DatabaseContentsError::class.java)
         }
     }
 
     @Test
-    fun `cannot add burden estimate set if cohort_size is missing from burden_outcome table`()
+    fun `cannot populate burden estimate set if cohort_size is missing from burden_outcome table`()
     {
         JooqContext().use { db ->
             setupDatabase(db)
             db.dsl.deleteFrom(BURDEN_OUTCOME).where(BURDEN_OUTCOME.CODE.eq("cohort_size")).execute()
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+                val setId = repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
+                repo.populateBurdenEstimateSet(setId, groupId, touchstoneId, scenarioId, data)
             }.isInstanceOf(DatabaseContentsError::class.java)
         }
     }
 
     @Test
-    fun `cannot add burden estimate set if responsibility set status is submitted`()
+    fun `cannot create burden estimate set if responsibility set status is submitted`()
     {
         JooqContext().use { db ->
             setupDatabase(db, responsibilitySetStatus = "submitted")
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+                repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
             }.isInstanceOf(OperationNotAllowedError::class.java)
                     .hasMessage("the following problems occurred:\nThe burden estimates uploaded for this touchstone have been submitted for review." +
                             " You cannot upload any new estimates.")
@@ -145,13 +166,13 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
     }
 
     @Test
-    fun `cannot add burden estimate set if responsibility set status is approved`()
+    fun `cannot create burden estimate set if responsibility set status is approved`()
     {
         JooqContext().use { db ->
             setupDatabase(db, responsibilitySetStatus = "approved")
             val repo = makeRepository(db)
             assertThatThrownBy {
-                repo.addBurdenEstimateSet(groupId, touchstoneId, scenarioId, data, username, timestamp)
+                repo.createBurdenEstimateSet(groupId, touchstoneId, scenarioId, username, timestamp)
             }.isInstanceOf(OperationNotAllowedError::class.java)
                     .hasMessage("the following problems occurred:\nThe burden estimates uploaded for this touchstone have been reviewed and approved." +
                             " You cannot upload any new estimates.")
@@ -159,19 +180,19 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
     }
 
     @Test
-    fun `cannot add burden estimate set if touchstone doesn't exist`()
+    fun `cannot create burden estimate set if touchstone doesn't exist`()
     {
         checkBadId(otherTouchstoneId = "wrong-id")
     }
 
     @Test
-    fun `cannot add burden estimate set if group doesn't exist`()
+    fun `cannot create burden estimate set if group doesn't exist`()
     {
         checkBadId(otherGroupId = "wrong-id")
     }
 
     @Test
-    fun `cannot add burden estimate set if scenario doesn't exist`()
+    fun `cannot create burden estimate set if scenario doesn't exist`()
     {
         checkBadId(otherScenarioId = "wrong-id")
     }
@@ -188,27 +209,6 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
                 repo.addBurdenEstimateSet(otherGroupId, otherTouchstoneId, otherScenarioId, data, username, timestamp)
             }.isInstanceOf(UnknownObjectError::class.java)
         }
-    }
-
-    private fun setupDatabase(db: JooqContext, addModel: Boolean = true,
-                              responsibilitySetStatus: String = "incomplete"): ReturnedIds
-    {
-        db.addTouchstone("touchstone", 1, "Touchstone 1", addName = true)
-        db.addScenarioDescription(scenarioId, "Test scenario", "Hib3", addDisease = true)
-        db.addGroup(groupId, "Test group")
-        val modelVersionId = if (addModel)
-        {
-            db.addModel(modelId, groupId, "Hib3")
-            db.addModelVersion(modelId, modelVersion, setCurrent = true)
-        }
-        else
-        {
-            null
-        }
-        val setId = db.addResponsibilitySet(groupId, touchstoneId, responsibilitySetStatus)
-        val responsibilityId = db.addResponsibility(setId, touchstoneId, scenarioId)
-        db.addUserForTesting(username)
-        return ReturnedIds(modelVersionId, responsibilityId)
     }
 
     private fun checkBurdenEstimates(db: JooqContext, setId: Int)
@@ -230,14 +230,18 @@ class BurdenEstimateRepositoryTestsDeprecated : RepositoryTests<BurdenEstimateRe
         checkRecord(records[5], setId, 1980, 30, "AGO", "deaths", 20.toDecimal())
     }
 
-    private fun checkBurdenEstimateSetMetadata(db: JooqContext, returnedIds: ReturnedIds): Int
+    private fun checkBurdenEstimateSetMetadata(db: JooqContext,
+                                               setId: Int,
+                                               returnedIds: ReturnedIds,
+                                               expectedStatus: String): Int
     {
         val t = BURDEN_ESTIMATE_SET
-        val set = db.dsl.fetchOne(t)
+        val set = db.dsl.selectFrom(t).where(t.ID.eq(setId)).fetchOne()
         assertThat(set[t.MODEL_VERSION]).isEqualTo(returnedIds.modelVersion!!)
         assertThat(set[t.RESPONSIBILITY]).isEqualTo(returnedIds.responsibility)
         assertThat(set[t.UPLOADED_BY]).isEqualTo(username)
         assertThat(set[t.UPLOADED_ON].toInstant()).isEqualTo(timestamp)
+        assertThat(set[t.STATUS]).isEqualTo(expectedStatus)
         return set[t.ID]
     }
